refactor: move blueprints to `react` package
