+++ .changeset/slimy-ghosts-rescue.md
@@ -1,0 +1,7 @@
+---
+'@backstage/plugin-catalog-unprocessed-entities': patch
+'@backstage/plugin-devtools-react': patch
+'@backstage/plugin-devtools': patch
+---
+
+Add support for adding `unprocessed-entities` and other tabs to `devtools`
