+++ plugins/devtools-react/README.md
@@ -1,0 +1,5 @@
+# @backstage/plugin-devtools-react
+
+Welcome to the web library package for the devtools plugin!
+
+_This plugin was created through the Backstage CLI_
