+++ plugins/devtools-react/catalog-info.yaml
@@ -1,0 +1,10 @@
+apiVersion: backstage.io/v1alpha1
+kind: Component
+metadata:
+  name: backstage-plugin-devtools-react
+  title: '@backstage/plugin-devtools-react'
+  description: Web library for the devtools plugin
+spec:
+  lifecycle: experimental
+  type: backstage-web-library
+  owner: maintainers
