+++ plugins/devtools-react/report.api.md
@@ -1,0 +1,54 @@
+## API Report File for "@backstage/plugin-devtools-react"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+import { ConfigurableExtensionDataRef } from '@backstage/frontend-plugin-api';
+import { ExtensionBlueprint } from '@backstage/frontend-plugin-api';
+import { ExtensionDataRef } from '@backstage/frontend-plugin-api';
+import { JSX as JSX_2 } from 'react';
+
+// @alpha
+export const DevToolsRouteBlueprint: ExtensionBlueprint<{
+  kind: 'devtools-route';
+  params: DevToolsRouteBlueprintParams;
+  output: ExtensionDataRef<DevToolsRouteData, 'devtools.route', {}>;
+  inputs: {};
+  config: {};
+  configInput: {};
+  dataRefs: {
+    route: ConfigurableExtensionDataRef<
+      DevToolsRouteData,
+      'devtools.route',
+      {}
+    >;
+  };
+}>;
+
+// @alpha
+export interface DevToolsRouteBlueprintParams {
+  // (undocumented)
+  loader: () => Promise<JSX_2.Element>;
+  // (undocumented)
+  path: string;
+  // (undocumented)
+  title: string;
+}
+
+// @alpha
+export interface DevToolsRouteData {
+  // (undocumented)
+  children: JSX_2.Element;
+  // (undocumented)
+  path: string;
+  // (undocumented)
+  title: string;
+}
+
+// @alpha
+export const devToolsRouteDataRef: ConfigurableExtensionDataRef<
+  DevToolsRouteData,
+  'devtools.route',
+  {}
+>;
+```
