+++ plugins/devtools-react/src/devToolsRouteDataRef.ts
