+++ plugins/devtools/src/alpha/index.ts
@@ -15,7 +15,3 @@
  */
 
 export { default } from './plugin';
-export { DevToolsRouteBlueprint } from './devToolsRouteBlueprint';
-export { devToolsRouteDataRef } from './devToolsRouteDataRef';
-export type { DevToolsRouteData } from './devToolsRouteDataRef';
-export type { DevToolsRouteBlueprintParams } from './devToolsRouteBlueprint';
